+++ src/lightning_fabric/CHANGELOG.md
@@ -30,6 +30,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Added support for managing callbacks via `Fabric(callbacks=...)` and emitting events through `Fabric.call()` ([#16074](https://github.com/Lightning-AI/lightning/issues/16074))
 
+- Added Logger support ([#16121](https://github.com/Lightning-AI/lightning/issues/16121))
+  * Added `Fabric(loggers=...)` to support different Logger frameworks in Fabric
+  * Added `Fabric.log` for logging scalars using multiple loggers
+  * Added `Fabric.log_dict` for logging a dictionary of multiple metrics at once
+  * Added `Fabric.loggers` and `Fabric.logger` attributes to access the individual logger instances
+
+
+- Added support for a consistent `.zero_grad(set_to_none=...)` on the wrapped optimizer regardless of which strategy is used ([#16275](https://github.com/Lightning-AI/lightning/issues/16275))
+
 
 ### Changed